ui: better fixture name, and corrected for linter warning
